Rework to best practices and SOC
