+++ Models/Dto/AlbumDto.cs
@@ -1,6 +1,4 @@
-﻿using PhotoGalleryAPI.Models.Data;
-
-namespace PhotoGalleryAPI.Models.Dto
+﻿namespace PhotoGalleryAPI.Models.Dto
 {
-    public record AlbumDto(string Name, User User, string ParentName, string Description);
+    public record AlbumDto(string Name, string ParentName, string Description);
 }
