+++ Services/IAuthService.cs
@@ -1,7 +1,7 @@
 ﻿namespace PhotoGalleryAPI.Services
 {
 
-    public interface IUserService
+    public interface IAuthService
     {
         void CreatePasswordHash(string password, out byte[] passwordHash, out byte[] passwordSalt);
         bool VerifyPasswordHash(string password, byte[] passwordHash, byte[] passwordSalt);
